déplacement des fonctions qui récupèrent les mapping
