+++ dags/sources/dags/source_pyreo.py
@@ -1,6 +1,6 @@
 from airflow import DAG
+from sources.config.airflow_params import get_mapping_config
 from sources.tasks.airflow_logic.operators import default_args, eo_task_chain
-from utils.base_utils import get_mapping_config
 
 with DAG(
     dag_id="eo-pyreo",
