+++ dags/utils/base_utils.py
@@ -1,9 +1,7 @@
 import csv
 import io
-import json
 import math
 import re
-from pathlib import Path
 from urllib.parse import urlparse
 
 import pandas as pd
@@ -39,39 +37,6 @@ def get_address(row, col="adresse_format_ban"):
     address, postal_code, city = extract_details(row, col)
 
     return pd.Series([address, postal_code, city])
-
-
-def get_mapping_config(mapping_key: str = "sous_categories"):
-    config_path = Path(__file__).parent.parent / "config" / "db_mapping.json"
-    with open(config_path, "r") as f:
-        config = json.load(f)
-    return config[mapping_key]
-
-
-def source_sinoe_dechet_mapping_get():
-    """Mapping de C_TYP_DECHET (ex: "01")
-    vers L_TYP_DECHET (ex: "Déchets de composés chimiques")
-    {
-    "total": 232,
-    "results": [
-        {
-        "_rand": 237210,
-        "C_TYP_DECHET": "01",
-        "_i": 1,
-        "NIV_HIER": 1,
-        "L_TYP_DECHET": "Déchets de composés chimiques",
-        "_score": null,
-        "_id": "4UbH7jVe1hc_lXPc0oWz9"
-        },
-    """
-    url = "https://data.ademe.fr/data-fair/api/v1/datasets/sinoe-r-nomenclature-dechets/lines?size=1000"
-    data = requests.get(url).json()
-    # Attention on a eu des renommage L_TYP_DECHET <-> LST_TYP_DECHET par le passé
-    # d'où le get sur les deux clés
-    return {
-        x["C_TYP_DECHET"]: x.get("L_TYP_DECHET", None) or x.get("LST_TYP_DECHET")
-        for x in data["results"]
-    }
 
 
 # TODO faire un retry avec tenacity
